feat: added `useOutsideClickListener()` hook
BREAKING CHANGE: default export replaced with named one
