+++ src/index.js
@@ -1,1 +1,7 @@
-module.exports = require('./OutsideClickListener');
+const OutsideClickListener = require('./OutsideClickListener');
+const useOutsideClickListener = require('./useOutsideClickListener');
+
+module.exports = {
+    OutsideClickListener,
+    useOutsideClickListener
+};
